% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{compile.cytof.model}
\alias{compile.cytof.model}
\title{Compile CyTOF model}
\usage{
compile.cytof.model(y_ls, K, L, data = model.data(y_ls),
  constants = model.consts(y = y_ls, K = K, L = L), inits = model.inits(y =
  y_ls), print_mcmc_config = FALSE, showCompilerOutput = FALSE)
}
\description{
TODO: write documentation
}
