% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_param}
\alias{get_param}
\title{Get column index of parameter}
\usage{
get_param(name, samples)
}
\arguments{
\item{name:}{name of parameter}

\item{samples:}{the samples object}
}
\description{
Get column index of parameter
}
