% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_constants.R
\name{model.consts}
\alias{model.consts}
\title{Generate Nimble model constants}
\usage{
model.consts(N, J, K, L, y, b_prior = solve_miss_mech_params(y = c(min(y,
  na.rm = TRUE), -0.1), p = c(0.99, 1e-04)))
}
\description{
TODO: document
}
