% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model.data}
\alias{model.data}
\title{Nimble model data object generator}
\usage{
model.data(y_ls)
}
\description{
y: concatenated matrix of transformed marker expression. missing values should be recorded as NA
}
