% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_zeros.R
\name{pad_zeros}
\alias{pad_zeros}
\title{Pad zeros to list of ragged matrices}
\usage{
pad_zeros(y_list)
}
\value{
a (non-ragged) 3D array
}
\description{
Pads each element of a list of (ragged) matrices with zeros so that they all
have the same dimensions. Ragged matrices are matrices with the same number
of columns but different number of rows.
}
\examples{
y_ls = list(matrix(1,2,5), matrix(2,3,5), matrix(3,1,5))
pad_zeros(y_ls)
}
