% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_mechanism.R
\name{solve_miss_mech_params}
\alias{solve_miss_mech_params}
\title{Solves for parameters in missing mechanism given two points on the mechanism}
\usage{
solve_miss_mech_params(y, p)
}
\description{
Solves for parameters in missing mechanism given two points on the mechanism
}
